* Work so far on #609
