+++ Source/Krypton Components/Krypton.Ribbon/Controls Ribbon/KryptonRibbon.cs
@@ -11,8 +11,6 @@
 #endregion
 
 
-using System.Linq;
-
 namespace Krypton.Ribbon
 {
     /// <summary>
