+++ Source/Krypton Components/Krypton.Ribbon/View Layout/ViewLayoutRibbonTabs.cs
@@ -11,8 +11,6 @@
 #endregion
 
 
-using System.Linq;
-
 namespace Krypton.Ribbon
 {
     /// <summary>
