+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Enumerations/PaletteEnumerations.cs
@@ -1,0 +1,252 @@
+﻿namespace Krypton.Toolkit
+{
+    #region Enumeration: SchemeOfficeColors
+
+    internal enum SchemeOfficeColors
+    {
+        TextLabelControl = 0,
+        TextButtonNormal = 1,
+        TextButtonChecked = 2,
+        ButtonNormalBorder = 3,
+        ButtonNormalDefaultBorder = 4,
+        ButtonNormalBack1 = 5,
+        ButtonNormalBack2 = 6,
+        ButtonNormalDefaultBack1 = 7,
+        ButtonNormalDefaultBack2 = 8,
+        ButtonNormalNavigatorBack1 = 9,
+        ButtonNormalNavigatorBack2 = 10,
+        PanelClient = 11,
+        PanelAlternative = 12,
+        ControlBorder = 13,
+        SeparatorHighBorder1 = 14,
+        SeparatorHighBorder2 = 15,
+        HeaderPrimaryBack1 = 16,
+        HeaderPrimaryBack2 = 17,
+        HeaderSecondaryBack1 = 18,
+        HeaderSecondaryBack2 = 19,
+        HeaderText = 20,
+        StatusStripText = 21,
+        ButtonBorder = 22,
+        SeparatorLight = 23,
+        SeparatorDark = 24,
+        GripLight = 25,
+        GripDark = 26,
+        ToolStripBack = 27,
+        StatusStripLight = 28,
+        StatusStripDark = 29,
+        ImageMargin = 30,
+        ToolStripBegin = 31,
+        ToolStripMiddle = 32,
+        ToolStripEnd = 33,
+        OverflowBegin = 34,
+        OverflowMiddle = 35,
+        OverflowEnd = 36,
+        ToolStripBorder = 37,
+        FormBorderActive = 38,
+        FormBorderInactive = 39,
+        FormBorderActiveLight = 40,
+        FormBorderActiveDark = 41,
+        FormBorderInactiveLight = 42,
+        FormBorderInactiveDark = 43,
+        FormBorderHeaderActive = 44,
+        FormBorderHeaderInactive = 45,
+        FormBorderHeaderActive1 = 46,
+        FormBorderHeaderActive2 = 47,
+        FormBorderHeaderInactive1 = 48,
+        FormBorderHeaderInactive2 = 49,
+        FormHeaderShortActive = 50,
+        FormHeaderShortInactive = 51,
+        FormHeaderLongActive = 52,
+        FormHeaderLongInactive = 53,
+        FormButtonBorderTrack = 54,
+        FormButtonBack1Track = 55,
+        FormButtonBack2Track = 56,
+        FormButtonBorderPressed = 57,
+        FormButtonBack1Pressed = 58,
+        FormButtonBack2Pressed = 59,
+        TextButtonFormNormal = 60,
+        TextButtonFormTracking = 61,
+        TextButtonFormPressed = 62,
+        LinkNotVisitedOverrideControl = 63,
+        LinkVisitedOverrideControl = 64,
+        LinkPressedOverrideControl = 65,
+        LinkNotVisitedOverridePanel = 66,
+        LinkVisitedOverridePanel = 67,
+        LinkPressedOverridePanel = 68,
+        TextLabelPanel = 69,
+        RibbonTabTextNormal = 70,
+        RibbonTabTextChecked = 71,
+        RibbonTabSelected1 = 72,
+        RibbonTabSelected2 = 73,
+        RibbonTabSelected3 = 74,
+        RibbonTabSelected4 = 75,
+        RibbonTabSelected5 = 76,
+        RibbonTabTracking1 = 77,
+        RibbonTabTracking2 = 78,
+        RibbonTabHighlight1 = 79,
+        RibbonTabHighlight2 = 80,
+        RibbonTabHighlight3 = 81,
+        RibbonTabHighlight4 = 82,
+        RibbonTabHighlight5 = 83,
+        RibbonTabSeparatorColor = 84,
+        RibbonGroupsArea1 = 85,
+        RibbonGroupsArea2 = 86,
+        RibbonGroupsArea3 = 87,
+        RibbonGroupsArea4 = 88,
+        RibbonGroupsArea5 = 89,
+        RibbonGroupBorder1 = 90,
+        RibbonGroupBorder2 = 91,
+        RibbonGroupTitle1 = 92,
+        RibbonGroupTitle2 = 93,
+        RibbonGroupBorderContext1 = 94,
+        RibbonGroupBorderContext2 = 95,
+        RibbonGroupTitleContext1 = 96,
+        RibbonGroupTitleContext2 = 97,
+        RibbonGroupDialogDark = 98,
+        RibbonGroupDialogLight = 99,
+        RibbonGroupTitleTracking1 = 100,
+        RibbonGroupTitleTracking2 = 101,
+        RibbonMinimizeBarDark = 102,
+        RibbonMinimizeBarLight = 103,
+        RibbonGroupCollapsedBorder1 = 104,
+        RibbonGroupCollapsedBorder2 = 105,
+        RibbonGroupCollapsedBorder3 = 106,
+        RibbonGroupCollapsedBorder4 = 107,
+        RibbonGroupCollapsedBack1 = 108,
+        RibbonGroupCollapsedBack2 = 109,
+        RibbonGroupCollapsedBack3 = 110,
+        RibbonGroupCollapsedBack4 = 111,
+        RibbonGroupCollapsedBorderT1 = 112,
+        RibbonGroupCollapsedBorderT2 = 113,
+        RibbonGroupCollapsedBorderT3 = 114,
+        RibbonGroupCollapsedBorderT4 = 115,
+        RibbonGroupCollapsedBackT1 = 116,
+        RibbonGroupCollapsedBackT2 = 117,
+        RibbonGroupCollapsedBackT3 = 118,
+        RibbonGroupCollapsedBackT4 = 119,
+        RibbonGroupFrameBorder1 = 120,
+        RibbonGroupFrameBorder2 = 121,
+        RibbonGroupFrameInside1 = 122,
+        RibbonGroupFrameInside2 = 123,
+        RibbonGroupFrameInside3 = 124,
+        RibbonGroupFrameInside4 = 125,
+        RibbonGroupCollapsedText = 126,
+        AlternatePressedBack1 = 127,
+        AlternatePressedBack2 = 128,
+        AlternatePressedBorder1 = 129,
+        AlternatePressedBorder2 = 130,
+        FormButtonBack1Checked = 131,
+        FormButtonBack2Checked = 132,
+        FormButtonBorderCheck = 133,
+        FormButtonBack1CheckTrack = 134,
+        FormButtonBack2CheckTrack = 135,
+        RibbonQATMini1 = 136,
+        RibbonQATMini2 = 137,
+        RibbonQATMini3 = 138,
+        RibbonQATMini4 = 139,
+        RibbonQATMini5 = 140,
+        RibbonQATMini1I = 141,
+        RibbonQATMini2I = 142,
+        RibbonQATMini3I = 143,
+        RibbonQATMini4I = 144,
+        RibbonQATMini5I = 145,
+        RibbonQATFullbar1 = 146,
+        RibbonQATFullbar2 = 147,
+        RibbonQATFullbar3 = 148,
+        RibbonQATButtonDark = 149,
+        RibbonQATButtonLight = 150,
+        RibbonQATOverflow1 = 151,
+        RibbonQATOverflow2 = 152,
+        RibbonGroupSeparatorDark = 153,
+        RibbonGroupSeparatorLight = 154,
+        ButtonClusterButtonBack1 = 155,
+        ButtonClusterButtonBack2 = 156,
+        ButtonClusterButtonBorder1 = 157,
+        ButtonClusterButtonBorder2 = 158,
+        NavigatorMiniBackColor = 159,
+        GridListNormal1 = 160,
+        GridListNormal2 = 161,
+        GridListPressed1 = 162,
+        GridListPressed2 = 163,
+        GridListSelected = 164,
+        GridSheetColNormal1 = 165,
+        GridSheetColNormal2 = 166,
+        GridSheetColPressed1 = 167,
+        GridSheetColPressed2 = 168,
+        GridSheetColSelected1 = 169,
+        GridSheetColSelected2 = 170,
+        GridSheetRowNormal = 171,
+        GridSheetRowPressed = 172,
+        GridSheetRowSelected = 173,
+        GridDataCellBorder = 174,
+        GridDataCellSelected = 175,
+        InputControlTextNormal = 176,
+        InputControlTextDisabled = 177,
+        InputControlBorderNormal = 178,
+        InputControlBorderDisabled = 179,
+        InputControlBackNormal = 180,
+        InputControlBackDisabled = 181,
+        InputControlBackInactive = 182,
+        InputDropDownNormal1 = 183,
+        InputDropDownNormal2 = 184,
+        InputDropDownDisabled1 = 185,
+        InputDropDownDisabled2 = 186,
+        ContextMenuHeadingBack = 187,
+        ContextMenuHeadingText = 188,
+        ContextMenuImageColumn = 189,
+        AppButtonBack1 = 190,
+        AppButtonBack2 = 191,
+        AppButtonBorder = 192,
+        AppButtonOuter1 = 193,
+        AppButtonOuter2 = 194,
+        AppButtonOuter3 = 195,
+        AppButtonInner1 = 196,
+        AppButtonInner2 = 197,
+        AppButtonMenuDocsBack = 198,
+        AppButtonMenuDocsText = 199,
+        SeparatorHighInternalBorder1 = 200,
+        SeparatorHighInternalBorder2 = 201,
+        RibbonGalleryBorder = 202,
+        RibbonGalleryBackNormal = 203,
+        RibbonGalleryBackTracking = 204,
+        RibbonGalleryBack1 = 205,
+        RibbonGalleryBack2 = 206,
+        RibbonTabTracking3 = 207,
+        RibbonTabTracking4 = 208,
+        RibbonGroupBorder3 = 209,
+        RibbonGroupBorder4 = 210,
+        RibbonGroupBorder5 = 211,
+        RibbonGroupTitleText = 212,
+        RibbonDropArrowLight = 213,
+        RibbonDropArrowDark = 214,
+        HeaderDockInactiveBack1 = 215,
+        HeaderDockInactiveBack2 = 216,
+        ButtonNavigatorBorder = 217,
+        ButtonNavigatorText = 218,
+        ButtonNavigatorTrack1 = 219,
+        ButtonNavigatorTrack2 = 220,
+        ButtonNavigatorPressed1 = 221,
+        ButtonNavigatorPressed2 = 222,
+        ButtonNavigatorChecked1 = 223,
+        ButtonNavigatorChecked2 = 224,
+        ToolTipBottom = 225,
+        // ============================================
+        MenuItemText = 226,
+        ContextMenuMargin = 227,
+        ContextMenuInner = 228,
+        ToolStripText = 229
+    }
+
+    #endregion
+
+    #region Enumeration: SchemeContextMenuColors
+    // Note: Add these to SchemeOfficeColors
+    internal enum SchemeContextMenuColors
+    {
+        MenuItemText = 1,
+        ContextMenuMargin = 2,
+        ContextMenuInner = 3,
+    }
+
+    #endregion
+}
