+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Office 2007/PaletteOffice2007BlueLightMode.cs
@@ -51,7 +51,7 @@ namespace Krypton.Toolkit
                                                                         Color.FromArgb(64, Color.White),    // Outside position
                                                                         Color.FromArgb(63, 101, 152)        // Border (normal) position
                                                                       };
-        private static readonly Color[] _schemeColors = { Color.FromArgb( 21,  66, 139),    // TextLabelControl
+        private static readonly Color[] _schemeOfficeColors = { Color.FromArgb( 21,  66, 139),    // TextLabelControl
                                                                       Color.FromArgb( 21,  66, 139),    // TextButtonNormal
                                                                       Color.Black,                      // TextButtonChecked
                                                                       Color.FromArgb(161, 189, 207),    // ButtonNormalBorder
@@ -297,7 +297,7 @@ namespace Krypton.Toolkit
         /// Initialize a new instance of the PaletteOffice2007Blue class.
         /// </summary>
         public PaletteOffice2007BlueLightMode()
-            : base(_schemeColors,
+            : base(_schemeOfficeColors,
                    _checkBoxList,
                    _galleryButtonList,
                    _radioButtonArray,
