+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Office 2010/PaletteOffice2010Black.cs
@@ -57,7 +57,7 @@ namespace Krypton.Toolkit
                                                                         Color.FromArgb(64, Color.White),    // Outside position
                                                                         Color.FromArgb(35, 35, 35)          // Border (normal) position
                                                                       };
-        private static readonly Color[] _schemeColors = { Color.FromArgb( 76,  83,  92),    // TextLabelControl
+        private static readonly Color[] _schemeOfficeColors = { Color.FromArgb( 76,  83,  92),    // TextLabelControl
                                                                       Color.Black,                      // TextButtonNormal
                                                                       Color.Black,                      // TextButtonChecked
                                                                       Color.FromArgb(106, 106, 106),    // ButtonNormalBorder1
@@ -318,7 +318,7 @@ namespace Krypton.Toolkit
         /// Initialize a new instance of the PaletteOffice2010Black class.
         /// </summary>
         public PaletteOffice2010Black()
-            : base(_schemeColors,
+            : base(_schemeOfficeColors,
                    _checkBoxList,
                    _galleryButtonList,
                    _radioButtonArray,
@@ -374,7 +374,7 @@ namespace Krypton.Toolkit
                     switch (state)
                     {
                         case PaletteState.Normal:
-                            return _schemeColors[(int)SchemeOfficeColors.HeaderSecondaryBack1];
+                            return _schemeOfficeColors[(int)SchemeOfficeColors.HeaderSecondaryBack1];
                     }
                     break;
             }
@@ -398,7 +398,7 @@ namespace Krypton.Toolkit
                     switch (state)
                     {
                         case PaletteState.Normal:
-                            return _schemeColors[(int)SchemeOfficeColors.ControlBorder];
+                            return _schemeOfficeColors[(int)SchemeOfficeColors.ControlBorder];
                     }
                     break;
             }
@@ -420,7 +420,7 @@ namespace Krypton.Toolkit
                     switch (state)
                     {
                         case PaletteState.Normal:
-                            return _schemeColors[(int)SchemeOfficeColors.ControlBorder];
+                            return _schemeOfficeColors[(int)SchemeOfficeColors.ControlBorder];
                     }
                     break;
             }
@@ -449,7 +449,7 @@ namespace Krypton.Toolkit
                 case PaletteContentStyle.ButtonCustom3:
                     if (state == PaletteState.NormalDefaultOverride)
                     {
-                        return _schemeColors[(int)SchemeOfficeColors.TextButtonChecked];
+                        return _schemeOfficeColors[(int)SchemeOfficeColors.TextButtonChecked];
                     }
 
                     break;
@@ -457,8 +457,8 @@ namespace Krypton.Toolkit
                 case PaletteContentStyle.ButtonNavigatorStack:
                 case PaletteContentStyle.ButtonNavigatorOverflow:
                     return state == PaletteState.NormalDefaultOverride
-                        ? _schemeColors[(int)SchemeOfficeColors.TextButtonChecked]
-                        : _schemeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
+                        ? _schemeOfficeColors[(int)SchemeOfficeColors.TextButtonChecked]
+                        : _schemeOfficeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
 
                 case PaletteContentStyle.HeaderPrimary:
                 case PaletteContentStyle.HeaderDockInactive:
@@ -494,8 +494,8 @@ namespace Krypton.Toolkit
                 case PaletteContentStyle.ButtonNavigatorStack:
                 case PaletteContentStyle.ButtonNavigatorOverflow:
                     return state == PaletteState.NormalDefaultOverride
-                        ? _schemeColors[(int)SchemeOfficeColors.TextButtonChecked]
-                        : _schemeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
+                        ? _schemeOfficeColors[(int)SchemeOfficeColors.TextButtonChecked]
+                        : _schemeOfficeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
 
                 case PaletteContentStyle.HeaderPrimary:
                 case PaletteContentStyle.HeaderDockInactive:
@@ -531,8 +531,8 @@ namespace Krypton.Toolkit
                 case PaletteContentStyle.ButtonNavigatorStack:
                 case PaletteContentStyle.ButtonNavigatorOverflow:
                     return state == PaletteState.NormalDefaultOverride
-                        ? _schemeColors[(int)SchemeOfficeColors.TextButtonChecked]
-                        : _schemeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
+                        ? _schemeOfficeColors[(int)SchemeOfficeColors.TextButtonChecked]
+                        : _schemeOfficeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
 
                 case PaletteContentStyle.HeaderPrimary:
                 case PaletteContentStyle.HeaderDockInactive:
@@ -568,8 +568,8 @@ namespace Krypton.Toolkit
                 case PaletteContentStyle.ButtonNavigatorStack:
                 case PaletteContentStyle.ButtonNavigatorOverflow:
                     return state == PaletteState.NormalDefaultOverride
-                        ? _schemeColors[(int)SchemeOfficeColors.TextButtonChecked]
-                        : _schemeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
+                        ? _schemeOfficeColors[(int)SchemeOfficeColors.TextButtonChecked]
+                        : _schemeOfficeColors[(int)SchemeOfficeColors.ButtonNavigatorText];
 
                 case PaletteContentStyle.HeaderPrimary:
                 case PaletteContentStyle.HeaderDockInactive:
