+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Office 2010/PaletteOffice2010White.cs
@@ -51,7 +51,7 @@ namespace Krypton.Toolkit
                                                                         Color.FromArgb(64, Color.White),    // Outside position
                                                                         Color.FromArgb(80, 81, 82)          // Border (normal) position
                                                                       };
-        private static readonly Color[] _schemeColors = new Color[] { Color.FromArgb( 59,  59,  59),    // TextLabelControl
+        private static readonly Color[] _schemeOfficeColors = new Color[] { Color.FromArgb( 59,  59,  59),    // TextLabelControl
                                                                       Color.FromArgb( 59,  59,  59),    // TextButtonNormal
                                                                       Color.Black,                      // TextButtonChecked
                                                                       Color.FromArgb(170, 170, 170),    // ButtonNormalBorder1 -n
@@ -310,7 +310,7 @@ namespace Krypton.Toolkit
         /// Initialize a new instance of the PaletteOffice2010Silver class.
         /// </summary>
         public PaletteOffice2010White()
-            : base(_schemeColors,
+            : base(_schemeOfficeColors,
                    _checkBoxList,
                    _galleryButtonList,
                    _radioButtonArray,
