+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Office 365/PaletteOffice365Blue.cs
@@ -48,7 +48,7 @@ namespace Krypton.Toolkit
                                                                         Color.FromArgb(64, Color.White),    // Outside position
                                                                         Color.FromArgb(63, 101, 152)        // Border (normal) position
                                                                       };
-        private static readonly Color[] _schemeColors = { Color.FromArgb( 30,  57,  91),    // TextLabelControl
+        private static readonly Color[] _schemeOfficeColors = { Color.FromArgb( 30,  57,  91),    // TextLabelControl
                                                                       Color.FromArgb( 30,  57,  91),    // TextButtonNormal
                                                                       Color.Black,                      // TextButtonChecked
                                                                       Color.FromArgb(171, 186, 208),    // ButtonNormalBorder
@@ -306,7 +306,7 @@ namespace Krypton.Toolkit
         /// <summary>
         /// Initializes a new instance of the PaletteOffice365Blue class.
         /// </summary>
-        public PaletteOffice365Blue() : base(_schemeColors, _checkBoxList, _galleryButtonList, _radioButtonArray, _trackBarColors)
+        public PaletteOffice365Blue() : base(_schemeOfficeColors, _checkBoxList, _galleryButtonList, _radioButtonArray, _trackBarColors)
         {
 
         }
